refactor: change Header -> SiteHeader
to prepare for AppHeader, which will share some commonality with
SiteHeader.

includes:
- extract Logo, SiteMenu, ProfileButton for reuse
- move Docs link into SiteDrawer and hide on smaller screens (because
there won't be as much space for it in the AppHeader)
