+++ src/web/common/components/Layout.tsx
@@ -2,7 +2,7 @@ import { NextPage } from "next";
 import { Roboto } from "next/font/google";
 import { ReactNode } from "react";
 
-import { Header } from "@/web/common/components/Header";
+import { SiteHeader } from "@/web/common/components/Header/SiteHeader";
 
 const roboto = Roboto({
   weight: ["300", "400", "500", "700"],
@@ -16,7 +16,7 @@ interface LayoutProps {
 const Layout: NextPage<LayoutProps> = ({ children }) => {
   return (
     <>
-      <Header />
+      <SiteHeader />
 
       {/* self-host font to prevent layout shift from fallback fonts loading first, see https://nextjs.org/docs/pages/building-your-application/optimizing/fonts */}
       <main className={roboto.className}>{children}</main>
